+++ lib/main.dart
@@ -6,8 +6,8 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:just_audio_background/just_audio_background.dart';
 import 'package:just_audio_media_kit/just_audio_media_kit.dart';
 import 'package:radio/app.dart';
-import 'package:radio/radio_station.dart';
-import 'package:radio/shared_pref.dart';
+import 'package:radio/models/radio_station.dart';
+import 'package:radio/utils/shared_pref.dart';
 import 'package:url_strategy/url_strategy.dart';
 
 Future<void> main() async {
