+++ lib/pages/privacy_policy_page.dart
