+++ lib/provider/favorite_stations.dart
@@ -1,5 +1,5 @@
+import 'package:radio/models/radio_station.dart';
 import 'package:radio/provider/stations.dart';
-import 'package:radio/radio_station.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 part 'favorite_stations.g.dart';
