+++ lib/provider/province_controller.dart
@@ -1,6 +1,6 @@
 import 'package:radio/models/province.dart';
+import 'package:radio/models/radio_station.dart';
 import 'package:radio/provider/stations.dart';
-import 'package:radio/radio_station.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 part 'province_controller.g.dart';
