+++ lib/provider/station_search.dart
@@ -1,7 +1,7 @@
 import 'package:fuzzy/fuzzy.dart';
+import 'package:radio/models/radio_station.dart';
 import 'package:radio/provider/search_input.dart';
 import 'package:radio/provider/stations.dart';
-import 'package:radio/radio_station.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 part 'station_search.g.dart';
